move ollama provider
