+++ crates/goose-providers/src/formats.rs
@@ -1,3 +1,4 @@
 pub mod anthropic;
+pub mod ollama;
 pub mod openai;
 pub mod openai_responses;
