+++ crates/goose-providers/src/formats/ollama.rs
@@ -10,20 +10,20 @@
 //! - qwen3-coder-32b
 
 use crate::conversation::message::{Message, MessageContent};
-use async_stream::try_stream;
-use chrono;
-use futures::Stream;
-use goose_providers::{
+use crate::{
     conversation::token_usage::ProviderUsage,
     formats::openai::{self, is_valid_function_name},
 };
+use async_stream::try_stream;
+use chrono;
+use futures::Stream;
 use regex::Regex;
 use rmcp::model::{object, CallToolRequestParams, ErrorCode, ErrorData, Role};
 use serde_json::Value;
 use std::borrow::Cow;
 use uuid::Uuid;
 
-pub use goose_providers::formats::openai::{
+pub use crate::formats::openai::{
     create_request, format_messages, format_tools, get_usage, validate_tool_schemas,
 };
