+++ crates/goose-providers/src/lib.rs
@@ -11,6 +11,7 @@ pub mod images;
 pub mod json;
 pub(crate) mod mcp_utils;
 pub mod model;
+pub mod ollama;
 pub mod openai;
 pub mod openai_compatible;
 pub mod permission;
