+++ crates/goose/src/providers/formats/mod.rs
@@ -6,6 +6,5 @@ pub mod bedrock;
 pub mod databricks;
 pub mod gcpvertexai;
 pub mod google;
-pub mod ollama;
 pub mod openrouter;
 pub mod snowflake;
